+++ docs/news.rst
@@ -89,6 +89,8 @@ Changes
   currently always returns the empty bytes value (``b""``).
 * ``ZstdDecompressionObj`` now has an ``unconsumed_tail`` attribute. This
   attribute currently always returns the empty bytes value (``b""``).
+* ``ZstdDecompressionObj`` now has an ``eof`` attribute returning whether the
+  compressed data has been fully read.
 
 0.17.0 (released 2021-01-18)
 ============================